refactor: 패키지명 수정 util -> constants
