+++ src/main/java/baseball/Player.java
@@ -1,7 +1,7 @@
 package baseball;
 
 import static camp.nextstep.edu.missionutils.Console.*;
-import static util.GameConstant.*;
+import static constants.GameConstant.*;
 
 import java.util.Arrays;
 import java.util.HashSet;
